requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
fluent-logger==0.9.3
idna==2.7
iso3166==0.9
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.15.3
msgpack==0.6.1
pkg-resources==0.0.0
pymongo==3.7.2
requests==2.20.0
six==1.11.0
urllib3==1.24
Werkzeug==0.14.1
WTForms==2.2.1
